Clean up navigation classes.
- Rename `Component` interface to `NavEntry`, remove where unnecessary, extract to new file and add documentation.
- Remove `StackNavigator` and `SlotNavigator` interfaces as they serve no purpose.
