+++ shared/feature/src/commonMain/kotlin/app/futured/kmptemplate/feature/navigation/root/RootSlotNavigator.kt
@@ -1,6 +1,5 @@
 package app.futured.kmptemplate.feature.navigation.root
 
-import app.futured.kmptemplate.util.arch.SlotNavigator
 import app.futured.kmptemplate.util.ext.asStateFlow
 import app.futured.kmptemplate.util.ext.componentCoroutineScope
 import com.arkivanov.decompose.ComponentContext
@@ -9,17 +8,15 @@ import com.arkivanov.decompose.router.slot.childSlot
 
 class RootSlotNavigator(
     private val slotNavigator: SlotNavigation<RootDestination>,
-) : SlotNavigator<RootDestination>, SlotNavigation<RootDestination> by slotNavigator {
+) : SlotNavigation<RootDestination> by slotNavigator {
 
-    fun createSlot(componentContext: ComponentContext) =
-        componentContext.childSlot(
-            source = slotNavigator,
-            serializer = RootDestination.serializer(),
-            initialConfiguration = { RootDestination.Login },
-            handleBackButton = false,
-            childFactory = { destination, childContext ->
-                destination.createComponent(childContext)
-            },
-        )
-            .asStateFlow(componentContext.componentCoroutineScope())
+    internal fun createSlot(componentContext: ComponentContext) = componentContext.childSlot(
+        source = slotNavigator,
+        serializer = RootDestination.serializer(),
+        initialConfiguration = { RootDestination.Login },
+        handleBackButton = false,
+        childFactory = { destination, childContext ->
+            destination.createComponent(childContext)
+        },
+    ).asStateFlow(componentContext.componentCoroutineScope())
 }
