+++ shared/feature/src/commonMain/kotlin/app/futured/kmptemplate/feature/ui/first/FirstScreen.kt
@@ -1,10 +1,9 @@
 package app.futured.kmptemplate.feature.ui.first
 
-import app.futured.kmptemplate.util.arch.Component
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
 
-interface FirstScreen : Component {
+interface FirstScreen {
     val viewState: StateFlow<FirstViewState>
     val actions: Actions
     val events: Flow<FirstUiEvent>
