+++ shared/feature/src/commonMain/kotlin/app/futured/kmptemplate/feature/ui/second/SecondScreen.kt
@@ -1,9 +1,8 @@
 package app.futured.kmptemplate.feature.ui.second
 
-import app.futured.kmptemplate.util.arch.Component
 import kotlinx.coroutines.flow.StateFlow
 
-interface SecondScreen : Component {
+interface SecondScreen {
     val viewState: StateFlow<SecondViewState>
     val actions: Actions
