+++ shared/feature/src/commonMain/kotlin/app/futured/kmptemplate/feature/ui/third/ThirdScreen.kt
@@ -1,9 +1,8 @@
 package app.futured.kmptemplate.feature.ui.third
 
-import app.futured.kmptemplate.util.arch.Component
 import kotlinx.coroutines.flow.StateFlow
 
-interface ThirdScreen : Component {
+interface ThirdScreen {
     val viewState: StateFlow<ThirdViewState>
     val actions: Actions
