+++ shared/util/src/commonMain/kotlin/app/futured/kmptemplate/util/arch/Navigation.kt
@@ -1,0 +1,16 @@
+package app.futured.kmptemplate.util.arch
+
+import com.arkivanov.decompose.ComponentContext
+
+/**
+ * This interface is used to mark a class as a navigation entry.
+ * This class usually wraps instance of a screen.
+ */
+interface NavEntry
+
+/**
+ * A base interface for destinations used in Decompose navigation.
+ */
+interface Destination<out C : NavEntry> {
+    fun createComponent(componentContext: ComponentContext): C
+}
